requirements3.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
awscli==1.15.19
boto3==1.7.14
botocore==1.10.19         # via awscli, boto3, s3transfer
colorama==0.3.7           # via awscli
docutils==0.14            # via awscli, botocore
futures==3.2.0 ; python_version == "2.7"
jmespath==0.9.4           # via boto3, botocore
pyasn1==0.4.5             # via rsa
python-dateutil==2.8.0    # via botocore
pyyaml==3.12              # via awscli
rsa==3.4.2                # via awscli
s3transfer==0.1.13        # via awscli, boto3
six==1.12.0               # via python-dateutil
